+++ examples/server.rs
@@ -1,3 +1,9 @@
+use std::{
+    sync::mpsc::{self, RecvTimeoutError},
+    thread,
+    time::Duration,
+};
+
 use open62541::{ua, ObjectNode, Server, VariableNode};
 use open62541_sys::{
     UA_NS0ID_BASEDATAVARIABLETYPE, UA_NS0ID_FOLDERTYPE, UA_NS0ID_OBJECTSFOLDER, UA_NS0ID_ORGANIZES,
@@ -7,7 +13,7 @@ use open62541_sys::{
 fn main() -> anyhow::Result<()> {
     env_logger::init();
 
-    let mut server = Server::new();
+    let (server, runner) = Server::new();
 
     println!("Adding server nodes");
 
@@ -36,11 +42,53 @@ fn main() -> anyhow::Result<()> {
 
     server.write_variable_string(&variable_node_id, "foobar")?;
 
-    println!("Running server");
+    let (cancel_tx, cancel_rx) = mpsc::channel();
 
-    server.run()?;
+    // Start background task that simulates changing variable values.
+    let server_task_handle = thread::spawn(move || -> anyhow::Result<()> {
+        println!("Simulating values");
+        loop {
+            for value in ["foo", "bar", "baz"] {
+                match cancel_rx.recv_timeout(Duration::from_millis(1000)) {
+                    Ok(()) => return Ok(()),
+                    Err(RecvTimeoutError::Timeout) => {
+                        // Continue and simulate next updated value below, then repeat loop.
+                    }
+                    Err(RecvTimeoutError::Disconnected) => panic!("main task should be running"),
+                }
+                server.write_variable_string(&variable_node_id, value)?;
+            }
+        }
+    });
+
+    // Start runner task that handles incoming connections (events).
+    let runner_task_handle = thread::spawn(|| -> anyhow::Result<()> {
+        println!("Running server");
+        runner.run()?;
+        Ok(())
+    });
+
+    // Wait for runner task to finish eventually (SIGINT/Ctrl+C).
+    if let Err(err) = runner_task_handle
+        .join()
+        .expect("runner task should not panic")
+    {
+        println!("Runner task failed: {err}");
+    }
 
     println!("Exiting");
+
+    cancel_tx.send(()).expect("server task should be running");
+
+    // Wait for simulation task to shut down after canceling.
+    if let Err(err) = server_task_handle
+        .join()
+        .expect("server task should not panic")
+    {
+        println!("Server task failed: {err}");
+    }
+
+    println!("Done");
 
     Ok(())
 }