+++ src/lib.rs
@@ -86,7 +86,7 @@ pub use self::{
     client::{Client, ClientBuilder},
     data_type::DataType,
     error::{Error, Result},
-    server::{Server, ServerBuilder},
+    server::{Server, ServerBuilder, ServerRunner},
     server_nodes::{ObjectNode, VariableNode},
     userdata::Userdata,
     value::{ScalarValue, ValueType, VariantValue},