+++ src/server.rs
@@ -1,4 +1,4 @@
-use std::{ffi::c_void, ptr};
+use std::{ffi::c_void, ptr, sync::Arc};
 
 use open62541_sys::{
     UA_ServerConfig, UA_Server_runUntilInterrupt, __UA_Server_addNode, __UA_Server_write,
@@ -58,8 +58,12 @@ impl ServerBuilder {
 
     /// Builds OPC UA server.
     #[must_use]
-    pub fn build(self) -> Server {
-        Server(ua::Server::new_with_config(self.0))
+    pub fn build(self) -> (Server, ServerRunner) {
+        let server = Arc::new(ua::Server::new_with_config(self.0));
+
+        let runner = ServerRunner(Arc::clone(&server));
+        let server = Server(server);
+        (server, runner)
     }
 
     /// Access server configuration.
@@ -71,9 +75,12 @@ impl ServerBuilder {
 
 /// OPC UA server.
 ///
-/// This represents an OPC UA server. Nodes can be added through the several methods below, and then
-/// the server can be started with [`run()`](Self::run).
-pub struct Server(ua::Server);
+/// This represents an OPC UA server. Nodes can be added through the several methods below.
+///
+/// Note: The server must be started with [`ServerRunner::run()`] before it can accept connections
+/// from clients.
+#[derive(Clone)]
+pub struct Server(Arc<ua::Server>);
 
 impl Server {
     /// Creates default server.
@@ -89,7 +96,7 @@ impl Server {
     ///
     /// See [`ServerBuilder::build()`].
     #[must_use]
-    pub fn new() -> Self {
+    pub fn new() -> (Self, ServerRunner) {
         ServerBuilder::default().build()
     }
 
@@ -98,10 +105,11 @@ impl Server {
     /// # Errors
     ///
     /// This fails when the node cannot be added.
-    pub fn add_object_node(&mut self, node: ObjectNode) -> Result<()> {
+    pub fn add_object_node(&self, node: ObjectNode) -> Result<()> {
         let status_code = ua::StatusCode::new(unsafe {
             __UA_Server_addNode(
-                self.0.as_mut_ptr(),
+                // SAFETY: Cast to `mut` pointer, function is marked `UA_THREADSAFE`.
+                self.0.as_ptr().cast_mut(),
                 // Passing ownership is trivial with primitive value (`u32`).
                 ua::NodeClass::OBJECT.into_raw(),
                 node.requested_new_node_id.as_ptr(),
@@ -124,10 +132,11 @@ impl Server {
     /// # Errors
     ///
     /// This fails when the node cannot be added.
-    pub fn add_variable_node(&mut self, node: VariableNode) -> Result<()> {
+    pub fn add_variable_node(&self, node: VariableNode) -> Result<()> {
         let status_code = ua::StatusCode::new(unsafe {
             __UA_Server_addNode(
-                self.0.as_mut_ptr(),
+                // SAFETY: Cast to `mut` pointer, function is marked `UA_THREADSAFE`.
+                self.0.as_ptr().cast_mut(),
                 // Passing ownership is trivial with primitive value (`u32`).
                 ua::NodeClass::VARIABLE.into_raw(),
                 node.requested_new_node_id.as_ptr(),
@@ -150,10 +159,11 @@ impl Server {
     /// # Errors
     ///
     /// This fails when the variable node cannot be written.
-    pub fn write_variable(&mut self, node_id: &ua::NodeId, value: &ua::Variant) -> Result<()> {
+    pub fn write_variable(&self, node_id: &ua::NodeId, value: &ua::Variant) -> Result<()> {
         let status_code = ua::StatusCode::new(unsafe {
             __UA_Server_write(
-                self.0.as_mut_ptr(),
+                // SAFETY: Cast to `mut` pointer, function is marked `UA_THREADSAFE`.
+                self.0.as_ptr().cast_mut(),
                 node_id.as_ptr(),
                 // Passing ownership is trivial with primitive value (`u32`).
                 ua::AttributeId::VALUE.into_raw(),
@@ -186,12 +196,17 @@ impl Server {
     /// # Errors
     ///
     /// This fails when the variable node cannot be written.
-    pub fn write_variable_string(&mut self, node_id: &ua::NodeId, value: &str) -> Result<()> {
+    pub fn write_variable_string(&self, node_id: &ua::NodeId, value: &str) -> Result<()> {
         let value = ua::String::new(value)?;
         let ua_variant = ua::Variant::init().with_scalar(&value);
         self.write_variable(node_id, &ua_variant)
     }
+}
 
+#[allow(clippy::module_name_repetitions)]
+pub struct ServerRunner(Arc<ua::Server>);
+
+impl ServerRunner {
     /// Runs the server until interrupted.
     ///
     /// The server is shut down cleanly upon receiving the `SIGINT` signal at which point the method
@@ -201,14 +216,15 @@ impl Server {
     ///
     /// This fails when the server cannot be started.
     pub fn run(self) -> Result<()> {
-        let status_code =
-            ua::StatusCode::new(unsafe { UA_Server_runUntilInterrupt(self.0.as_mut_ptr()) });
+        let status_code = ua::StatusCode::new(unsafe {
+            UA_Server_runUntilInterrupt(
+                // SAFETY: Cast to `mut` pointer. Function is not marked `UA_THREADSAFE` but we make
+                // sure that it can only be invoked a single time (ownership of `ServerRunner`). The
+                // examples in `open62541` demonstrate that running the server in its own thread and
+                // interacting with it as we do through `Server` is okay.
+                self.0.as_ptr().cast_mut(),
+            )
+        });
         Error::verify_good(&status_code)
-    }
-}
-
-impl Default for Server {
-    fn default() -> Self {
-        Self::new()
     }
 }