+++ src/ua/server.rs
@@ -10,6 +10,14 @@ use crate::{ua, Error};
 /// [`UA_Server_delete()`].
 pub struct Server(NonNull<UA_Server>);
 
+// SAFETY: We know that the underlying `UA_Server` allows access from different threads, i.e. it may
+// be dropped in a different thread from where it was created.
+unsafe impl Send for Server {}
+
+// SAFETY: The underlying `UA_Server` can be used from different threads concurrently, at least with
+// _most_ methods (those marked `UA_THREADSAFE` and/or with explicit mutex locks inside).
+unsafe impl Sync for Server {}
+
 impl Server {
     /// Creates server from server config.
     ///
@@ -42,7 +50,7 @@ impl Server {
     /// The value is owned by `Self`. Ownership must not be given away, in whole or in parts. This
     /// may happen when `open62541` functions are called that take ownership of values by pointer.
     #[must_use]
-    pub(crate) const unsafe fn as_mut_ptr(&self) -> *mut UA_Server {
+    pub(crate) unsafe fn as_mut_ptr(&mut self) -> *mut UA_Server {
         self.0.as_ptr()
     }
 }